+++ x/wasm/alias.go
@@ -93,10 +93,12 @@ type (
 	MsgStoreCode            = types.MsgStoreCode
 	MsgInstantiateContract  = types.MsgInstantiateContract
 	MsgExecuteContract      = types.MsgExecuteContract
+	MsgMigrateContract      = types.MsgMigrateContract
+	MsgUpdateAdministrator  = types.MsgUpdateAdministrator
 	Model                   = types.Model
 	CodeInfo                = types.CodeInfo
 	ContractInfo            = types.ContractInfo
-	CreatedAt               = types.CreatedAt
+	CreatedAt               = types.AbsoluteTxPosition
 	WasmConfig              = types.WasmConfig
 	MessageHandler          = keeper.MessageHandler
 	BankEncoder             = keeper.BankEncoder